requirements.txt
torch>=2.2.0
transformers>=4.40.0
pandas>=2.0.0
numpy>=1.26.0
scikit-learn>=1.4.0
ta>=0.11.0
snownlp>=0.12.3
akshare>=1.12.0
requests>=2.31.0
pyyaml>=6.0.1
matplotlib>=3.8.0
mplfinance>=0.12.10b0
tqdm>=4.66.0
